clock

+++ player/player.py
@@ -673,7 +673,7 @@ class Player(object):
             radar = radar_screenshot(factor=0.5)
             radar = radar.rotate(90, expand=True)
             width, height = radar.size
-            bg.paste(radar, (600-width, 228+height/2))
+            bg.paste(radar, (int(600-width), int(228+height/2)))
 
         self.pimoroni.set_image(bg, saturation=PIMORONI_SATURATION)
         self.pimoroni.show(busy_wait=False)
